safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A10_T8-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Parse error
S7.9_A10_T8

PASS Expected parsing failure

TEST COMPLETE
